refactor: 解耦 VideoPlayer
