+++ src/renderer/src/components/VideoPlayer/VideoPlayer.tsx
@@ -1,28 +1,30 @@
-import React, { useState, useRef, useEffect, useCallback } from 'react'
+import React, { useRef, useCallback } from 'react'
 import ReactPlayer from 'react-player'
 import { VideoPlaceholder } from './VideoPlaceholder'
 import { LoadingIndicator } from '../LoadingIndicator'
 import { ErrorIndicator } from '../ErrorIndicator'
 import { VideoControlsFullScreen } from './VideoControlsFullScreen'
-import { useFullscreenMode } from '@renderer/hooks/useFullscreenMode'
 import {
   useVideoPlayerRef,
   useVideoPlayState,
-  useVideoError,
-  useVideoControls
+  useVideoError
 } from '@renderer/hooks/useVideoPlayerHooks'
 import { usePlayingVideoContext } from '@renderer/hooks/usePlayingVideoContext'
 // 导入新的统一控制器
 import { useReactPlayerController } from '@renderer/hooks/useReactPlayerController'
+// 导入内聚的功能 hooks / Import cohesive functionality hooks
+import { useVideoControlsDisplay } from '@renderer/hooks/useVideoControlsDisplay'
+import { useVideoPlayerInteractions } from '@renderer/hooks/useVideoPlayerInteractions'
+import { useVideoTextSelection } from '@renderer/hooks/useVideoTextSelection'
+import { useVideoPlayerNotifications } from '@renderer/hooks/useVideoPlayerNotifications'
 
 // 导入样式
 import styles from './VideoPlayer.module.css'
 import RendererLogger from '@renderer/utils/logger'
 import { SubtitleOverlay } from '@renderer/components/VideoPlayer/SubtitleOverlay'
 import { useVideoConfig } from '@renderer/hooks/useVideoConfig'
-import { useSubtitleCopy } from '@renderer/hooks/useSubtitleCopy'
-import { useCopySuccessToast } from '@renderer/hooks/useCopySuccessToast'
 import { CopySuccessToast } from '@renderer/components/CopySuccessToast/CopySuccessToast'
+import { useFullscreenMode } from '@renderer/hooks/useFullscreenMode'
 
 interface VideoPlayerProps {
   isVideoLoaded: boolean
@@ -43,12 +45,23 @@ function VideoPlayer({
   const playerRef = useVideoPlayerRef()
   const isPlaying = useVideoPlayState()
   const videoError = useVideoError()
+  const { isFullscreen } = useFullscreenMode()
 
   // 获取播放设置（用于不需要响应变化的逻辑）
   const { playbackRate, volume } = useVideoConfig()
 
-  // 获取控制方法
-  const { toggle } = useVideoControls()
+  // 内聚功能 hooks / Cohesive functionality hooks
+  const controlsDisplay = useVideoControlsDisplay()
+  const playerInteractions = useVideoPlayerInteractions({
+    showControlsWithTimeout: controlsDisplay.showControlsWithTimeout
+  })
+  const textSelection = useVideoTextSelection()
+
+  // 外部通知管理
+  useVideoPlayerNotifications({ onFullscreenToggle })
+
+  // 节流相关的 refs（保留在组件中，因为它们是实现细节）
+  const mouseMoveThrottleTimeoutRef = useRef<NodeJS.Timeout | null>(null)
 
   RendererLogger.componentRender({
     component: 'VideoPlayer',
@@ -60,68 +73,6 @@ function VideoPlayer({
     }
   })
 
-  // 内部状态管理 - 只管理 UI 相关的本地状态
-  const [showControls, setShowControls] = useState(false)
-  const [isPausedByHover, setIsPausedByHover] = useState(false)
-  const hideControlsTimeoutRef = useRef<NodeJS.Timeout | null>(null)
-
-  // 优化：添加节流相关的 refs
-  const lastMouseMoveTimeRef = useRef(0)
-  const mouseMoveThrottleTimeoutRef = useRef<NodeJS.Timeout | null>(null)
-
-  // 全屏状态管理
-  const { isFullscreen } = useFullscreenMode()
-
-  // 选中文本状态 / Selected text state
-  const [selectedText, setSelectedText] = useState<string>('')
-
-  // 复制成功提示管理 / Copy success toast management
-  const { toastState, showCopySuccess, hideCopySuccess } = useCopySuccessToast()
-
-  // 字幕复制功能 / Subtitle copy functionality
-  useSubtitleCopy({
-    selectedText,
-    enabled: true,
-    onCopySuccess: showCopySuccess
-  })
-
-  // 监听全屏状态变化并通知父组件
-  useEffect(() => {
-    onFullscreenToggle?.(isFullscreen)
-  }, [isFullscreen, onFullscreenToggle])
-
-  // 字幕相关的回调函数
-  const handleWordHoverForControls = useCallback((isHovering: boolean) => {
-    if (isHovering) {
-      setShowControls(true)
-    }
-  }, [])
-
-  // 划词选中功能的回调 / Text selection callback
-  const handleSelectionChange = useCallback((selectedText: string) => {
-    setSelectedText(selectedText)
-    if (selectedText) {
-      console.log('选中的文本 / Selected text:', selectedText)
-      // 这里可以添加更多的处理逻辑，比如显示翻译等
-      // Additional logic can be added here, such as showing translation, etc.
-    }
-  }, [])
-
-  const handlePauseOnHover = useCallback(() => {
-    if (isPlaying) {
-      toggle()
-      setIsPausedByHover(true)
-    }
-  }, [isPlaying, toggle])
-
-  // 处理鼠标离开单词时恢复播放 / Handle resuming playback when mouse leaves word
-  const handleResumeOnLeave = useCallback(() => {
-    if (isPausedByHover) {
-      toggle()
-      setIsPausedByHover(false)
-    }
-  }, [isPausedByHover, toggle])
-
   // 使用新控制器提供的事件处理器
   const eventHandlers = playerController.createEventHandlers()
 
@@ -132,103 +83,22 @@ function VideoPlayer({
     eventHandlers.onReady()
   }, [eventHandlers, onVideoReady])
 
-  // 优化：提取控制栏显示逻辑，避免重复代码
-  const showControlsWithTimeout = useCallback((timeout: number = 3000) => {
-    setShowControls(true)
-
-    if (hideControlsTimeoutRef.current) {
-      clearTimeout(hideControlsTimeoutRef.current)
-    }
-
-    hideControlsTimeoutRef.current = setTimeout(() => {
-      setShowControls(false)
-    }, timeout)
-  }, [])
-
-  // 智能控制显示逻辑
-  const handleMouseEnter = useCallback((): void => {
-    showControlsWithTimeout()
-  }, [showControlsWithTimeout])
-
-  // 鼠标离开时隐藏控制栏
-  const handleMouseLeave = useCallback((): void => {
-    hideControlsTimeoutRef.current = setTimeout(() => {
-      setShowControls(false)
-    }, 2000) // 2秒后隐藏
-  }, [])
-
-  // 优化：添加节流的鼠标移动处理
-  const handleMouseMove = useCallback((): void => {
-    const now = Date.now()
-    const timeSinceLastMove = now - lastMouseMoveTimeRef.current
-
-    // 如果已经显示控制栏且时间间隔小于100ms，则跳过更新
-    if (showControls && timeSinceLastMove < 100) {
-      return
-    }
-
-    lastMouseMoveTimeRef.current = now
-
-    // 只有在控制栏未显示时才更新状态
-    if (!showControls) {
-      setShowControls(true)
-    }
-
-    // 重置定时器
-    if (hideControlsTimeoutRef.current) {
-      clearTimeout(hideControlsTimeoutRef.current)
-    }
-
-    hideControlsTimeoutRef.current = setTimeout(() => {
-      setShowControls(false)
-    }, 3000) // 3秒无操作后隐藏
-  }, [showControls])
-
-  // 处理视频播放器点击事件
-  const handleVideoClick = useCallback((): void => {
-    toggle()
-    showControlsWithTimeout()
-  }, [toggle, showControlsWithTimeout])
-
-  // 清理定时器
-  useEffect(() => {
+  // 清理定时器（保留在组件中）
+  React.useEffect(() => {
     return () => {
-      if (hideControlsTimeoutRef.current) {
-        clearTimeout(hideControlsTimeoutRef.current)
-      }
       if (mouseMoveThrottleTimeoutRef.current) {
-        // eslint-disable-next-line react-hooks/exhaustive-deps
         clearTimeout(mouseMoveThrottleTimeoutRef.current)
       }
     }
   }, [])
 
-  // 播放状态变化时的控制逻辑
-  useEffect(() => {
-    if (!isPlaying) {
-      setShowControls(true)
-      if (hideControlsTimeoutRef.current) {
-        clearTimeout(hideControlsTimeoutRef.current)
-      }
-    } else {
-      hideControlsTimeoutRef.current = setTimeout(() => {
-        setShowControls(false)
-      }, 3000)
-    }
-
-    // 如果是因为悬停暂停的，当播放状态变为播放时，重置悬停状态
-    if (isPlaying && isPausedByHover) {
-      setIsPausedByHover(false)
-    }
-  }, [isPlaying, isPausedByHover])
-
   return (
     <div className={styles.videoSection}>
       <div
         className={styles.videoContainer}
-        onMouseEnter={handleMouseEnter}
-        onMouseLeave={handleMouseLeave}
-        onMouseMove={handleMouseMove}
+        onMouseEnter={controlsDisplay.handleMouseEnter}
+        onMouseLeave={controlsDisplay.handleMouseLeave}
+        onMouseMove={controlsDisplay.handleMouseMove}
       >
         {playingVideoContext.videoFile ? (
           <>
@@ -246,7 +116,7 @@ function VideoPlayer({
               onReady={handleReactPlayerReady}
               onError={eventHandlers.onError}
               onLoadStart={eventHandlers.onLoadStart}
-              onClick={handleVideoClick}
+              onClick={playerInteractions.handleVideoClick}
               controls={false}
               progressInterval={300}
               style={{ cursor: 'pointer' }}
@@ -274,13 +144,11 @@ function VideoPlayer({
             {/* 错误状态提示 */}
             {videoError && <ErrorIndicator error={videoError} />}
 
-            {/* 字幕显示组件 - 独立组件，不会导致 VideoPlayer 频繁渲染 */}
+            {/* 字幕显示组件 - 使用内聚的回调 */}
             <SubtitleOverlay
-              onWordHover={handleWordHoverForControls}
-              onPauseOnHover={handlePauseOnHover}
-              onResumeOnLeave={handleResumeOnLeave}
+              onWordHover={controlsDisplay.handleWordHoverForControls}
               enableTextSelection={true}
-              onSelectionChange={handleSelectionChange}
+              onSelectionChange={textSelection.handleSelectionChange}
             />
 
             {/* 全屏控制栏 */}
@@ -288,7 +156,7 @@ function VideoPlayer({
               <VideoControlsFullScreen
                 isVideoLoaded={isVideoLoaded}
                 videoError={videoError}
-                showControls={showControls}
+                showControls={controlsDisplay.showControls}
               />
             )}
           </>
@@ -297,12 +165,12 @@ function VideoPlayer({
         )}
       </div>
 
-      {/* 复制成功提示 / Copy success toast */}
+      {/* 复制成功提示 - 使用内聚的文本选择功能 */}
       <CopySuccessToast
-        visible={toastState.visible}
-        position={toastState.position}
-        copiedText={toastState.copiedText}
-        onComplete={hideCopySuccess}
+        visible={textSelection.toastState.visible}
+        position={textSelection.toastState.position}
+        copiedText={textSelection.toastState.copiedText}
+        onComplete={textSelection.hideCopySuccess}
       />
     </div>
   )
