+++ src/renderer/src/components/VideoPlayer/index.ts
@@ -1,0 +1,23 @@
+/**
+ * VideoPlayer组件导出索引
+ * VideoPlayer components export index
+ */
+
+// 字幕文本组件 / Subtitle text components
+export {
+  OriginalSubtitleText,
+  ChineseSubtitleText,
+  EnglishSubtitleText,
+  type SubtitleTextProps
+} from './SubtitleText'
+
+// 双语字幕组件 / Bilingual subtitle components
+export { BilingualSubtitleLine, type BilingualSubtitleLineProps } from './BilingualSubtitle'
+
+// 字幕占位符组件 / Subtitle placeholder component
+export { SubtitlePlaceholder, type SubtitlePlaceholderProps } from './SubtitlePlaceholder'
+
+// 基础组件 / Base components
+export { WordWrapper, type WordWrapperProps } from './WordWrapper'
+export { TextRenderer, type TextRendererProps } from './TextRenderer'
+export { SmartTextContent, type SmartTextContentProps } from './SmartTextContent'
